requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.2
channels==2.4.0
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
isort==5.4.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
